ml_model/requirements.txt
# Core ML + YOLO
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python-headless>=4.8.0.74
pillow>=10.0.0
scikit-image>=0.21.0

# Math & utilities
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.1
shapely>=2.0.0

# Optional (development & testing)
matplotlib>=3.7.1

